2_Programas_informaticos/05_INFERENCIA_EN_LA_PLACA/code/requirements.txt
#numpy==1.19.5
numpy
setuptools
scipy
#scikit-learn==0.20.0
#scikit-image
# opencv-python== 4.4.0.42
# opencv-contrib-python== 4.4.0.46 
# tensorflow-gpu==1.15.2
# keras_applications
# tensorflow-model-optimization==0.5.0
# tensorflow-addons
matplotlib
#matplotlib ==3.6
tqdm
pillow
# mnn
Cython
pycocotools
# keras2onnx
# tf2onnx
# onnx
# onnxruntime
# tfcoreml
sympy
imgaug
imagecorruptions
bokeh
tidecv
